Move node creation out with test
