+++ gatsby-node.js
@@ -10,6 +10,7 @@ const {
     validateEnvVariables,
 } = require('./src/utils/setup/validate-env-variables');
 const { onCreatePage } = require('./src/utils/setup/on-create-page');
+const { createAssetNodes } = require('./src/utils/setup/create-asset-nodes');
 const { createTagPageType } = require('./src/utils/setup/create-tag-page-type');
 const { getMetadata } = require('./src/utils/get-metadata');
 const {
@@ -57,23 +58,7 @@ exports.sourceNodes = async ({
     }
 
     documents.forEach(doc => {
-        const pageNode = getNestedValue(['ast', 'children'], doc);
-        if (pageNode) {
-            // Cache static assets to save file i/o on repeated builds
-            doc.static_assets.forEach(asset => {
-                const assetNode = {
-                    id: asset,
-                    parent: null,
-                    children: [],
-                    internal: {
-                        type: 'Asset',
-                        content: asset,
-                        contentDigest: createContentDigest(asset),
-                    },
-                };
-                createNode(assetNode);
-            });
-        }
+        createAssetNodes(doc, createNode, createContentDigest);
         // Mimics onCreateNode
         postprocessDocument(doc, PAGE_ID_PREFIX, pages, slugContentMapping);
     });
